refactor(Page): Extract version toast
